+++ src/pages/[pageName]/[pageSlug].tsx
@@ -55,6 +55,7 @@ import {
   isValidPageSlug,
   isValidPageName,
   executeFunctionWithTimeout,
+  setSSRCacheHeaders,
 } from '../../libs/server/page'
 import {
   transformArticleStream,
@@ -137,6 +138,7 @@ export const getServerSideProps: GetServerSideProps = wrapper.getServerSideProps
             req,
           }
           log(options)
+          setSSRCacheHeaders(res)
           return {
             props: {
               menuItems,